Move ThumbnailData
